Replace + with , in civetweb port string.
Civetweb uses , to supply multiple parameters to options.  Ceph uses ,
to separate options.  This means we can't supply a multi-valued option
to civetweb.  But we want the civetweb listening_ports parameter to be
multi-valued, because that way we can have one civetweb instance
listen to both ssl and non-ssl ports.  By changing + to , in
the ports string, we can avoid the use of , in the ceph configuration
file and still give civetweb what it wants to see as a separator.

Signed-off-by: Marcus Watts <mwatts@redhat.com>